+++ packages/components/CHANGELOG.md
@@ -41,7 +41,7 @@
 ### Enhancement
 
 -   `QueryControls`: Add menu_order sorting option if supported by the post type. ([#68781](https://github.com/WordPress/gutenberg/pull/68781)).
--   `Popover`: Introduce `overflowOffset` prop to support adding an offset between popover and viewport edge ([#69555](https://github.com/WordPress/gutenberg/pull/69555)).
+-   `Popover`: Introduce `overflowOffsetY` prop to support adding a vertical offset between popover and viewport edge ([#69555](https://github.com/WordPress/gutenberg/pull/69555)).
 
 ### Internal